pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "windmill-agent-cli"
version = "0.1.0"
description = "Agent-native CLI for Windmill workspaces using the HTTP API directly."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "httpx>=0.27,<1.0",
  "PyYAML>=6.0,<7.0",
  "typer>=0.12,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
]

[project.scripts]
wmx = "wmx.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/wmx"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
